Cambio en Estructura de datos, comentarios y Modularización
